+++ android/src/main/java/org/openforis/collect/android/util/MessageSources.java
@@ -12,7 +12,7 @@ public abstract  class MessageSources {
 
     public static String getMessage(MessageSource messageSource, String key, Locale locale, Locale defaultLocale) {
         String message = messageSource.getMessage(locale, key);
-        if (message == null && defaultLocale != null && !locale.equals(defaultLocale)) {
+        if (message == null && defaultLocale != null && !defaultLocale.equals(locale)) {
             message = messageSource.getMessage(defaultLocale, key);
         }
         return message;