+++ CHANGELOG.md
@@ -5,8 +5,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [1.7.3] - build 160 2019-04-18
+### Fixed
+- Record creation and modified dates in data export
+
+## [1.7.3] - build 158 2019-04-16
+### Fixed
+- Language labels in Settings not showing in some devices
+
+## [1.7.3] - build 157 2019-04-16
+### Added
+- About page
+
+### Fixed
+- "One-time" calculated attribute export issue
+
 ## [1.7.2] - build 156 2019-04-11
 ### Added
 - Support Survey Guide files (if attached to the survey)
+
 ### Fixed
 - Error importing survey on certain devices